+++ app/src/main/java/ch/abwesend/privatecontacts/application/KoinModule.kt
@@ -59,7 +59,7 @@ import ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts.ser
 import ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts.service.AndroidContactSaveService
 import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.ContactToVCardMapper
 import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.VCardToContactMapper
-import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata.ToPhysicalAddressMapper
+import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata.import.ToPhysicalAddressMapper
 import ch.abwesend.privatecontacts.infrastructure.repository.vcard.repository.VCardImportExportRepository
 import ch.abwesend.privatecontacts.infrastructure.repository.vcard.repository.VCardRepository
 import ch.abwesend.privatecontacts.infrastructure.room.database.AppDatabase
