+++ app/src/main/java/ch/abwesend/privatecontacts/infrastructure/repository/vcard/mapping/VCardToContactMapper.kt
@@ -19,9 +19,9 @@ import ch.abwesend.privatecontacts.domain.model.result.generic.ErrorResult
 import ch.abwesend.privatecontacts.domain.model.result.generic.SuccessResult
 import ch.abwesend.privatecontacts.domain.util.Constants
 import ch.abwesend.privatecontacts.domain.util.injectAnywhere
-import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata.ToPhysicalAddressMapper
-import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata.toContactData
-import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata.toContactGroup
+import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata.import.ToPhysicalAddressMapper
+import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata.import.toContactData
+import ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata.import.toContactGroup
 import ezvcard.VCard
 
 // TODO add unit tests
