+++ app/src/main/java/ch/abwesend/privatecontacts/infrastructure/repository/vcard/mapping/contactdata/import/ToPhoneNumberMapper.kt
@@ -4,7 +4,7 @@
  * Florian Gubler
  */
 
-package ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata
+package ch.abwesend.privatecontacts.infrastructure.repository.vcard.mapping.contactdata.import
 
 import ch.abwesend.privatecontacts.domain.model.contactdata.ContactData
 import ch.abwesend.privatecontacts.domain.model.contactdata.ContactDataType
